move generation for pawns, rook, bishop and queen! with en passant and promotion. All pseudo legal of course ;)

+++ src/Piece.cpp
@@ -19,6 +19,6 @@ PieceList::PieceList() {
 
 std::string PieceType2Str(PieceType type) {
 	std::string tmp = "";
-	tmp += types.at(type);
+	tmp += types.at((int)type);
 	return tmp;
 }
